requirements.txt
Pillow     9.4.0
pip        22.3.1
setuptools 65.5.0